+++ crates/cli/src/app.rs
@@ -1,9 +1,8 @@
 use crate::commands::{
-    debug::DebugConfigArgs,
     plugin::{AddPluginArgs, InfoPluginArgs, ListPluginsArgs, RemovePluginArgs, SearchPluginArgs},
     ActivateArgs, AliasArgs, BinArgs, CleanArgs, CompletionsArgs, DiagnoseArgs, InstallArgs,
-    ListArgs, ListRemoteArgs, MigrateArgs, OutdatedArgs, PinArgs, RegenArgs, RunArgs, SetupArgs,
-    StatusArgs, UnaliasArgs, UninstallArgs, UnpinArgs, UpgradeArgs,
+    MigrateArgs, OutdatedArgs, PinArgs, RegenArgs, RunArgs, SetupArgs, StatusArgs, UnaliasArgs,
+    UninstallArgs, UnpinArgs, UpgradeArgs, VersionsArgs,
 };
 use clap::builder::styling::{Color, Style, Styles};
 use clap::{Parser, Subcommand, ValueEnum};
@@ -80,6 +79,7 @@ fn create_styles() -> Styles {
 )]
 pub struct App {
     #[arg(
+        value_enum,
         long,
         short = 'c',
         global = true,
@@ -98,12 +98,30 @@ pub struct App {
 
     #[arg(
         value_enum,
+        default_value_t,
         long,
         global = true,
         env = "PROTO_LOG",
         help = "Lowest log level to output"
     )]
-    pub log: Option<LogLevel>,
+    pub log: LogLevel,
+
+    #[arg(
+        long,
+        short = 'y',
+        global = true,
+        env = "PROTO_YES",
+        help = "Avoid all interactive prompts and use defaults"
+    )]
+    pub yes: bool,
+
+    #[arg(
+        long,
+        global = true,
+        env = "PROTO_JSON",
+        help = "Print as JSON (when applicable)"
+    )]
+    pub json: bool,
 
     #[command(subcommand)]
     pub command: Commands,
@@ -111,18 +129,16 @@ pub struct App {
 
 impl App {
     pub fn setup_env_vars(&self) {
-        let version = env!("CARGO_PKG_VERSION");
-
-        if let Some(level) = &self.log {
-            env::set_var("PROTO_APP_LOG", level.to_string());
-        } else if let Ok(level) = env::var("PROTO_LOG") {
-            env::set_var("PROTO_APP_LOG", level);
-        }
-
-        env::set_var("PROTO_VERSION", version);
+        env::set_var("PROTO_APP_LOG", self.log.to_string());
+        env::set_var("PROTO_VERSION", env!("CARGO_PKG_VERSION"));
 
         if let Ok(value) = env::var("PROTO_DEBUG_COMMAND") {
             env::set_var("WARPGATE_DEBUG_COMMAND", value);
+        }
+
+        // Disable ANSI colors in JSON output
+        if self.json {
+            env::set_var("NO_COLOR", "1");
         }
     }
 }
@@ -183,21 +199,6 @@ pub enum Commands {
         long_about = "Download and install one or many tools by version into ~/.proto/tools.\n\nIf no arguments are provided, will install all tools configured in .prototools.\n\nIf a name argument is provided, will install a single tool by version."
     )]
     Install(InstallArgs),
-
-    #[command(
-        alias = "ls",
-        name = "list",
-        about = "List installed versions for a tool."
-    )]
-    List(ListArgs),
-
-    #[command(
-        alias = "lsr",
-        name = "list-remote",
-        about = "List available versions for a tool.",
-        long_about = "List available versions by resolving versions from the tool's remote release manifest."
-    )]
-    ListRemote(ListRemoteArgs),
 
     #[command(
         name = "migrate",
@@ -277,6 +278,14 @@ pub enum Commands {
         about = "Upgrade proto to the latest version."
     )]
     Upgrade(UpgradeArgs),
+
+    #[command(
+        alias = "vs",
+        name = "versions",
+        about = "List available versions for a tool.",
+        long_about = "List available versions for a tool by resolving versions from the tool's remote release manifest."
+    )]
+    Versions(VersionsArgs),
 }
 
 #[derive(Clone, Debug, Subcommand)]
@@ -285,7 +294,7 @@ pub enum DebugCommands {
         name = "config",
         about = "Debug all loaded .prototools config's for the current directory."
     )]
-    Config(DebugConfigArgs),
+    Config,
 
     #[command(name = "env", about = "Debug the current proto environment and store.")]
     Env,
@@ -296,7 +305,7 @@ pub enum PluginCommands {
     #[command(
         name = "add",
         about = "Add a plugin to manage a tool.",
-        long_about = "Add a plugin to the local ./.prototools config, or global ~/.proto/.prototools config."
+        long_about = "Add a plugin to a .prototools config file to enable and manage that tool."
     )]
     Add(AddPluginArgs),
 
@@ -315,7 +324,7 @@ pub enum PluginCommands {
     #[command(
         name = "remove",
         about = "Remove a plugin and unmanage a tool.",
-        long_about = "Remove a plugin from the local ./.prototools config, or global ~/.proto/.prototools config."
+        long_about = "Remove a plugin from a .prototools config file and unmanage that tool."
     )]
     Remove(RemovePluginArgs),