+++ crates/cli/src/commands/unalias.rs
@@ -1,9 +1,9 @@
-use crate::helpers::{map_pin_type, PinOption};
 use crate::session::ProtoSession;
 use clap::Args;
-use proto_core::{Id, ProtoConfig};
+use iocraft::element;
+use proto_core::{Id, PinLocation, ProtoConfig};
 use starbase::AppResult;
-use starbase_styles::color;
+use starbase_console::ui::*;
 
 #[derive(Args, Clone, Debug)]
 pub struct UnaliasArgs {
@@ -13,15 +13,8 @@ pub struct UnaliasArgs {
     #[arg(required = true, help = "Alias name")]
     alias: String,
 
-    #[arg(
-        long,
-        group = "pin",
-        help = "Remove from the global ~/.proto/.prototools"
-    )]
-    global: bool,
-
-    #[arg(long, group = "pin", help = "Location of .prototools to remove from")]
-    from: Option<PinOption>,
+    #[arg(long, default_value_t, help = "Location of .prototools to remove from")]
+    from: PinLocation,
 }
 
 #[tracing::instrument(skip_all)]
@@ -29,36 +22,46 @@ pub async fn unalias(session: ProtoSession, args: UnaliasArgs) -> AppResult {
     let tool = session.load_tool(&args.id).await?;
     let mut value = None;
 
-    let config_path = ProtoConfig::update(
-        tool.proto
-            .get_config_dir(map_pin_type(args.global, args.from)),
-        |config| {
-            if let Some(tool_configs) = &mut config.tools {
-                if let Some(tool_config) = tool_configs.get_mut(&tool.id) {
-                    if let Some(aliases) = &mut tool_config.aliases {
-                        value = aliases.remove(&args.alias);
-                    }
+    let config_path = ProtoConfig::update(tool.proto.get_config_dir(args.from), |config| {
+        if let Some(tool_configs) = &mut config.tools {
+            if let Some(tool_config) = tool_configs.get_mut(&tool.id) {
+                if let Some(aliases) = &mut tool_config.aliases {
+                    value = aliases.remove(&args.alias);
                 }
             }
-        },
-    )?;
+        }
+    })?;
 
     let Some(value) = value else {
-        eprintln!(
-            "Alias {} not found in config {}",
-            color::id(&args.alias),
-            color::path(config_path)
-        );
+        session.console.render(element! {
+            Notice(variant: Variant::Caution) {
+                StyledText(
+                    content: format!(
+                        "Alias <label>{}</label> for <id>{}</id> not found in config <path>{}</path>",
+                        args.alias,
+                        args.id,
+                        config_path.display()
+                    ),
+                )
+            }
+        })?;
 
         return Ok(Some(1));
     };
 
-    println!(
-        "Removed alias {} ({}) from config {}",
-        color::id(&args.alias),
-        color::muted_light(value.to_string()),
-        color::path(config_path)
-    );
+    session.console.render(element! {
+        Notice(variant: Variant::Success) {
+            StyledText(
+                content: format!(
+                    "Removed <id>{}</id> alias <label>{}</label> <mutedlight>({})</mutedlight> from config <path>{}</path>",
+                    args.id,
+                    args.alias,
+                    value.to_string(),
+                    config_path.display()
+                ),
+            )
+        }
+    })?;
 
     Ok(None)
 }