+++ crates/cli/src/helpers.rs
@@ -1,42 +1,14 @@
-use clap::ValueEnum;
 use dialoguer::{
     console::{style, Style},
     theme::ColorfulTheme,
 };
 use indicatif::{MultiProgress, ProgressBar, ProgressDrawTarget, ProgressStyle};
 use miette::IntoDiagnostic;
-use proto_core::PinType;
+use semver::Version;
 use starbase_styles::color::{self, Color};
 use starbase_utils::env::bool_var;
-use std::{io::IsTerminal, time::Duration};
+use std::io::IsTerminal;
 use tracing::debug;
-
-#[derive(Clone, Copy, Debug, Default, ValueEnum)]
-pub enum PinOption {
-    #[value(alias = "store")]
-    Global,
-    #[default]
-    #[value(alias = "cwd")]
-    Local,
-    #[value(alias = "home")]
-    User,
-}
-
-pub fn map_pin_type(global: bool, option: Option<PinOption>) -> PinType {
-    if let Some(option) = option {
-        return match option {
-            PinOption::Global => PinType::Global,
-            PinOption::Local => PinType::Local,
-            PinOption::User => PinType::User,
-        };
-    }
-
-    if global {
-        PinType::Global
-    } else {
-        PinType::Local
-    }
-}
 
 pub fn create_theme() -> ColorfulTheme {
     ColorfulTheme {
@@ -100,21 +72,25 @@ pub fn create_progress_bar_download_style() -> ProgressStyle {
         .unwrap()
 }
 
-pub fn create_progress_spinner_style() -> ProgressStyle {
-    let mut chars = vec![];
+pub fn create_progress_loader_frames() -> Vec<String> {
+    let mut frames = vec![];
 
     for i in 1..=20 {
         if i == 20 {
-            chars.push("━".repeat(20));
+            frames.push("━".repeat(20));
         } else {
-            chars.push(format!("{}╾{}", "━".repeat(i - 1), " ".repeat(20 - i)));
+            frames.push(format!("{}╾{}", "━".repeat(i - 1), " ".repeat(20 - i)));
         }
     }
 
-    let chars = chars.iter().map(|c| c.as_str()).collect::<Vec<_>>();
+    frames
+}
+pub fn create_progress_spinner_style() -> ProgressStyle {
+    let frames = create_progress_loader_frames();
+    let frames = frames.iter().map(|f| f.as_str()).collect::<Vec<_>>();
 
     ProgressStyle::default_spinner()
-        .tick_strings(&chars)
+        .tick_strings(&frames)
         .template(format_template_styles("{prefix} {spinner:20.183/239} | {msg}").as_str())
         .unwrap()
 }
@@ -147,21 +123,6 @@ pub fn create_progress_bar<S: AsRef<str>>(start: S) -> ProgressBar {
     pb
 }
 
-pub fn create_progress_spinner<S: AsRef<str>>(start: S) -> ProgressBar {
-    let pb = if is_hidden_progress() {
-        ProgressBar::hidden()
-    } else {
-        ProgressBar::new_spinner()
-    };
-
-    pb.set_style(create_progress_spinner_style());
-    pb.enable_steady_tick(Duration::from_millis(100));
-
-    print_progress_state(&pb, start.as_ref().to_owned());
-
-    pb
-}
-
 // When not a TTY, we should display something to the user!
 pub fn print_progress_state(pb: &ProgressBar, message: String) {
     if message.is_empty() || pb.message() == message {
@@ -176,7 +137,7 @@ pub fn print_progress_state(pb: &ProgressBar, message: String) {
     }
 }
 
-pub async fn fetch_latest_version() -> miette::Result<String> {
+pub async fn fetch_latest_version() -> miette::Result<Version> {
     let version = reqwest::get("https://raw.githubusercontent.com/moonrepo/proto/master/version")
         .await
         .into_diagnostic()?
@@ -188,5 +149,5 @@ pub async fn fetch_latest_version() -> miette::Result<String> {
 
     debug!("Found latest version {}", color::hash(&version));
 
-    Ok(version)
+    Ok(Version::parse(&version).unwrap())
 }