+++ crates/cli/tests/clean_test.rs
@@ -1,5 +1,7 @@
 mod utils;
 
+use std::fs;
+use std::time::{Duration, SystemTime};
 use utils::*;
 
 mod clean {
@@ -17,96 +19,31 @@ mod clean {
     }
 
     #[test]
-    fn purges_tool_inventory() {
+    fn cleans_plugins() {
         let sandbox = create_empty_proto_sandbox();
-        sandbox.create_file(".proto/tools/node/1.2.3/index.js", "");
-        sandbox.create_file(".proto/tools/node/4.5.6/index.js", "");
+        sandbox.create_file(".proto/plugins/node_plugin.wasm", "{}");
+        sandbox.create_file(".proto/plugins/npm_plugin.wasm", "{}");
+
+        fs::File::options()
+            .write(true)
+            .open(sandbox.path().join(".proto/plugins/node_plugin.wasm"))
+            .unwrap()
+            .set_times(
+                fs::FileTimes::new().set_accessed(
+                    SystemTime::now()
+                        .checked_sub(Duration::from_secs(86400 * 2))
+                        .unwrap(),
+                ),
+            )
+            .unwrap();
 
         sandbox
             .run_bin(|cmd| {
-                cmd.arg("clean").arg("--yes").arg("--purge").arg("node");
-            })
-            .success();
-
-        assert!(!sandbox
-            .path()
-            .join(".proto/tools/node/1.2.3/index.js")
-            .exists());
-        assert!(!sandbox
-            .path()
-            .join(".proto/tools/node/4.5.6/index.js")
-            .exists());
-    }
-
-    #[cfg(not(windows))]
-    #[test]
-    fn purges_tool_bin() {
-        let sandbox = create_empty_proto_sandbox();
-        sandbox.create_file(".proto/tools/node/1.2.3/fake/file", "");
-        sandbox.create_file(
-            ".proto/tools/node/manifest.json",
-            r#"{ "installed_versions": ["1.2.3"] }"#,
-        );
-        sandbox.create_file(".proto/bin/other", "");
-
-        let bin1 = sandbox.path().join(".proto/bin/node");
-        let bin2 = sandbox.path().join(".proto/bin/node-1");
-        let bin3 = sandbox.path().join(".proto/bin/node-1.2");
-        let src = sandbox.path().join(".proto/tools/node/1.2.3/fake/file");
-
-        #[allow(deprecated)]
-        std::fs::soft_link(&src, &bin1).unwrap();
-        #[allow(deprecated)]
-        std::fs::soft_link(&src, &bin2).unwrap();
-        #[allow(deprecated)]
-        std::fs::soft_link(&src, &bin3).unwrap();
-
-        sandbox
-            .run_bin(|cmd| {
-                cmd.arg("clean").arg("--yes").arg("--purge").arg("node");
-            })
-            .success();
-
-        assert!(!bin1.exists());
-        assert!(bin1.symlink_metadata().is_err());
-        assert!(!bin2.exists());
-        assert!(bin2.symlink_metadata().is_err());
-        assert!(!bin3.exists());
-        assert!(bin3.symlink_metadata().is_err());
-    }
-
-    #[test]
-    fn purges_tool_shims() {
-        let sandbox = create_empty_proto_sandbox();
-        sandbox.create_file(".proto/shims/npm", "");
-        sandbox.create_file(".proto/shims/npm.exe", "");
-        sandbox.create_file(".proto/shims/npx", "");
-        sandbox.create_file(".proto/shims/npx.exe", "");
-
-        sandbox
-            .run_bin(|cmd| {
-                cmd.arg("clean").arg("--yes").arg("--purge").arg("npm");
-            })
-            .success();
-
-        if cfg!(windows) {
-            assert!(!sandbox.path().join(".proto/shims/npm.exe").exists());
-            assert!(!sandbox.path().join(".proto/shims/npx.exe").exists());
-        } else {
-            assert!(!sandbox.path().join(".proto/shims/npm").exists());
-            assert!(!sandbox.path().join(".proto/shims/npx").exists());
-        }
-    }
-
-    #[test]
-    fn purges_plugins() {
-        let sandbox = create_empty_proto_sandbox();
-        sandbox.create_file(".proto/plugins/node_plugin.wasm", "");
-        sandbox.create_file(".proto/plugins/npm_plugin.wasm", "");
-
-        sandbox
-            .run_bin(|cmd| {
-                cmd.arg("clean").arg("--yes").arg("--purge-plugins");
+                cmd.arg("clean")
+                    .arg("--yes")
+                    .arg("plugins")
+                    .arg("--days")
+                    .arg("1");
             })
             .success();
 
@@ -114,7 +51,7 @@ mod clean {
             .path()
             .join(".proto/plugins/node_plugin.wasm")
             .exists());
-        assert!(!sandbox
+        assert!(sandbox
             .path()
             .join(".proto/plugins/npm_plugin.wasm")
             .exists());