+++ crates/cli/tests/status_test.rs
@@ -77,6 +77,8 @@ mod status {
                 .current_dir(sandbox.path().join("a/b/c"));
         });
 
+        assert.debug();
+
         let output = assert.output();
 
         assert!(predicate::str::contains("node").eval(&output));