+++ CHANGELOG.md
@@ -10,6 +10,23 @@
 - [Rust](https://github.com/moonrepo/tools/blob/master/tools/rust/CHANGELOG.md)
 - [Schema (TOML, JSON, YAML)](https://github.com/moonrepo/tools/blob/master/tools/internal-schema/CHANGELOG.md)
 
+## Unreleased
+
+#### 💥 Breaking
+
+- Removed the `proto list` and `proto list-remote` commands, use `proto versions` instead.
+- Removed the `--global` option from `proto alias`, `unalias`, `pin`, and `unpin`, use `--to` or `--from` instead.
+- Removed the `--purge` option from `proto clean`, use `proto uninstall` instead.
+- Removed the `--purge-plugins` option from `proto clean`, use `proto clean plugins` instead.
+
+#### 🚀 Updates
+
+- Added a `--yes` option to `proto outdated`, that skips confirmation prompts.
+- Added a `--json` option to `proto clean`, to capture the cleaned result as JSON.
+- Added a new command, `proto versions <tool>`, that lists all available remote and installed versions/aliases.
+- Updated `proto clean` to accept a target in which to clean as the 1st argument. For example, `proto clean cache`.
+- Moved the `--json` and `--yes` options into global options.
+
 ## 0.43.3
 
 #### 🐞 Fixes
@@ -107,6 +124,10 @@
 - WASM API
   - Deprecated `LocateExecutablesOutput.primary` and `LocateExecutablesOutput.secondary` (use `exes` instead).
   - Updated `ToolMetadataOutput.plugin_version` to a `Version` type instead of `String`.
+
+#### 💥 Breaking
+
+- Removed `--global` arg from `proto plugin add|remove`.
 
 #### 🚀 Updates